--- a/Dockerfile
+++ b/Dockerfile
@@ -12,6 +12,8 @@ FROM ghcr.io/astral-sh/uv:python3.12-bookworm-slim
         RUN python -c 'import streamlit, dhlab, pandas, requests'
         RUN timeout 5s streamlit run emneord.py; exit 0
 
-        CMD streamlit run emneord.py --server.port ${PORT} --server.baseUrlPath /emneord-test-fast
-        CMD streamlit run emneord.py --server.port ${PORT} --server.baseUrlPath /emneord
+        CMD streamlit run emneord.py \
+            --server.port ${PORT} \
+            --browser.gatherUsageStats=False \
+            --server.baseUrlPath /emneord