--- a/Dockerfile
+++ b/Dockerfile
@@ -10,7 +10,7 @@ FROM ghcr.io/astral-sh/uv:python3.12-bookworm-slim
 
         # Warm up caches
         RUN python -c 'import streamlit, dhlab, pandas, requests'
-        RUN timeout 5s streamlit run emneord.py; exit 0
+        RUN timeout 5s streamlit hello; exit 0
 
         CMD streamlit run emneord.py \
             --server.port ${PORT} \