### Summary
* Use slim docker image
  - Faster container launch and build
* Use uv instead of pip
  - Faster builds
* Warm up python and streamlit caches by pre-running `python -c "import ..."` and `streamlit run hello`
  - Faster streamlit server launch and import loading
* Lazy-load imports
  - Much faster UI loading

### Dockerfile optimization samples from dhlab-app-dokumentstatistikk:
Total importtime after UI load, before lazy imports:
* 0.535s - `--compile-bytecode` only (default behavior in non-uv pip)
* 0.286s - `--compile-bytecode` and `streamlit run document_statistics.py`
* 0.288s - `--compile-bytecode` and `streamlit run document_statistics.py` + `import ...`

Total importtime after UI load AND lazy imports:
* 2.434s - `--compile-bytecode` only (default behavior in non-uv pip)
* 2.155s - `--compile-bytecode` + `streamlit run document_statistics.py`
* 1.761s - `--compile-bytecode` + `streamlit run document_statistics.py` + `import ...`